Rename DynamicTableScan to DynamicSeqScan, for consistency.
DynamicTableScan was the only one that didn't follow the naming scheme,
where the dynamic scan node has the same name as the normal version, with
the Dynamic prefix. We now have a consistent naming scheme:

SeqScan and DynamicSeqScan
IndexScan and DynamicIndexScan
BitmapHeapScan and DynamicBitmapHeapScan
BitmapIndexScan and DynamicBitmapIndexScan

Some fields, variables and functions still have DynamicTable as part of
the name, but in those cases, DynamicTable doesn't refer to the
DynamicSeqScan node type, but to all Dynamic scan variants. DynamicTable
seems appropriate for that use.

To avoid user-visible changes, EXPLAIN still prints "Dynamic Table Scan".
I actually think we should change that, too, but let's discuss that
separately.

Reviewed-by: Ashwin Agrawal <aagrawal@pivotal.io>
Reviewed-by: Bhuvnesh Chaudhary <bchaudhary@pivotal.io>