Reset synchronous_standby_names on FTS mirror promotion.
When FTS promotes mirror, in gp_segment_configuration it marks mirror as
not in sync ('n'). Hence, on mirror when promotion request is received,
best to reset the synchronous_standby_names. If this is not done,
commits hang/wait after promotion till next FTS probe cycle to detect
this condition and reset the synchronous_standby_names.