Remove unused rpm spec file
It has been a long time since this spec file has been used to build a
Greenplum rpm installer. The ability to build a Greenplum rpm installer
will eventually be replaced with a new spec file, but it will not live
in this directory nor does it need to be based on this older spec file.

Authored-by: Kris Macoskey <kmacoskey@pivotal.io>