Remove 4.3 specific pg_upgrade test data
Back when the initial backport of pg_upgrade was added to Greenplum
we included a small dump of a 4.3.x cluster for testing purposes.
This cluster was inadequate for testing pg_upgrade back then, and is
even more so today. Remove to reduce potential confusion that these
files are actually part of the now quite extensive test harness for
pg_upgrade in Greenplum. The cluster also requires access to a 4.3
cluster which is a licensed product, so remove from the open source
repository.

Discussion: https://github.com/greenplum-db/gpdb/pull/6149
Reviewed-by: Jim Doty <jdoty@pivotal.io>
Reviewed-by: Jacob Champion <pchampion@pivotal.io>