Fix for scan output costing
When we cost a scan, the cost of actually producing the output data is added to
the parent node in the query tree. See method CCostModelGPDB::CostChildren().
There is a special case in this method for a filter on top of a scan.

In this case, the cost of producing the output is based on the number of rows produced
by the filter, not the full scan. This code has some issues when we have a
filter on top of a partition selector on top of a scan. In this case, the cost
of producing the outputs is based on the full table scan and that cost is added
to the cost of the partition selector.

Our fix:

In method CCostModelGPDB::CostChildren(), recognize the pattern
filter(partition_selector(scan)) and cost it appropriately.

Co-authored-by: Sambitesh Dash <sdash@pivotal.io>