Do not use hardcoded TemplateDbOid for database access during standby promotion. (#6601)
TemplateDbOid is for database template1, but template1 could be recreated and
thus its oid is not longer TemplateDbOid.

Also change to Use database postgres to find database oid and tablespace like
what fts and gdd do. We should better use the same database for such
purpose. I thought about which database gpdb should use: template1 or
postgres. Both of them are template database, but it seems that some users
or customers sometimes customize their own template1 database and
template1 is the default template for database creating so if some auxiliary
processes use that, "create database" command will fail. Using database
postgres seems to be more reasonable. In addition, of course, users possibly
drop the database postgres, but that is a rare case and they could easily
recreate one for our purpose. We really do not need to over-design for such rare case.