Refactor executor Squelch and EagerFree functions.
Merge the two concepts, squelching and eager-freeing. There is now only
one function, ExecSquelchNode(), that you can call. It recurses to
children, as before, but it now also performs eager-freeing of resources.
Previously that was done as a separate call, but that was an unnecessary
distinction, because all callers of ExecSquelchNode() also called
ExecEagerFree()

The concept of eager-freeing still lives on, as ExecEagerFree*() functions
specific to many node types. But it no longer recurses! The pattern is
that ExecSquelchNode() always performs eager freeing of the node, and also
recurses. In addition to that, some node types also call the node-specific
EagerFree function of the same node, after reaching the end of tuples.

This makes it more clear which function should be called when.

ExecEagerWalker() used to have special handling for the pattern of:
  Result
  -> Material
    -> Broadcast Motion

I tried removing that, but then I started to get "illegal rescan of motion
node" errors in the regression tests, from queries with a LIMIT node in a
subquery. Upon closer look, I believe that was because the Limit node was
calling ExecSquelchNode() on the input, even though the Limit node was
marked as rescannable. To fix that, I added delayEagerFree logic to Limit
node, to not call ExecSquelchNode() when the node might get rescanned
later.

The planstate_walk_kids() code did not know how to recurse into the
children of a MergeAppend node. We missed adding that logic, when we
merged the MergeAppend node type from upstream, in 9.1. We don't use that
mechanism for recursing in ExecSquelchNode() anymore, but that probably
should be fixed, anyway, as a separate commit later.

Fixes https://github.com/greenplum-db/gpdb/issues/6602 and
https://github.com/greenplum-db/gpdb/issues/6074.

Reviewed-by: Tang Pengzhou <ptang@pivotal.io>