pg_isready: handle PQPING_MIRROR_READY and reassign exported constant
The GPDB-specific constant PQPING_MIRROR_READY, which indicates that a
mirror is ready for replication, was not handled in pg_isready.

Additionally, the value we selected for PQPING_MIRROR_READY might at one
point in the future conflict with upstream libpq, which would be a pain
to untangle. Try to avoid that situation by increasing the value.

Co-authored-by: Shoaib Lari <slari@pivotal.io>