Use centos6-test image for test jobs
and centos6-build for the OSS build job. The image for this job was left
unmodified when we previously introduced the centos6-build image for
enterprise compile jobs. We changed the OSS job to also use the
centos6-build image so we could remove the centos-gpdb-test-6 image.

We also renamed the image resources to
centos<version>-<test,build>-gpdb6, to clarify what they are
building/testing. Docker resource names should match the repository and
tag of the docker image.

Co-authored-by: David Sharp <dsharp@pivotal.io>
Co-authored-by: Ben Christel <bchristel@pivotal.io>
Co-authored-by: Karen Huddleston <khuddleston@pivotal.io>