Fix distributed snapshot xmax check.
As part of commit dc78e56, logic for distributed snapshot was modified
to use latestCompletedDxid. This changed the logic from xmax being
inclusive range to not inclusive for visible transactions in snapshot.
Hence, updating the check to return
DISTRIBUTEDSNAPSHOT_COMMITTED_INPROGRESS even for transaction id equal
to global xmax now. Other way to fix is using latestCompletedDxid
without +1 for xmax, but better is to keep logic similar to local
snapshot check and not have xmax in inclusive range of visible
transactions.

This was exposed in CI by test
isolation/results/heap-repeatable-read-vacuum-freeze failing
intermittently.  This was due to isolation framework itself triggering
query on pg_locks to check for deadlocks. This commit adds explicitely
test to cover the scenario.

Co-authored-by: Ashwin Agrawal <aagrawal@pivotal.io>