Make isolation2 result sets look more like psql's.
Why, you might ask? The next commit will modify the code in gpdiff.pl, so
that it doesn't get fooled by "----"-style comments, thinking that they
are psql result sets. A side-effect of that is that it would also no
longer recognize the result sets in the isolation2 output, without this
patch.