Detect beginning of result set better in gpdiff.
Improve the detection of the beginning of a result set. Previously, it
would get confused by comments like "-------", which look a lot like the
beginning of a single-column psql result set. That doesn't matter much,
as long as the test is passing, but if such a test fails, the diff was
very difficult to read, as atmsort reordered the SQL lines, too.

Make the detection more resilient, by looking at the previous line. In a
real psql result set, the previous line should be a header line, like
" col1 | col2 ". A header line begins and ends with spaces, anything else
means that we're seeing a SQL comment rather than a psql result set.

While we're at it, if the "------" line has any leading or trailing
whitespace, it's not a psql result set. I'm not sure why we were lenient
on that, but let's make that more strict, too.

Reviewed-by: Asim R P <apraveen@pivotal.io>