Fix regression failure on "gpmapreduce --help".
The test was using "-- ignore", to cause gpdiff to ignore any differences
in the test output. But after commit 1f44603af2, gpdiff doesn't consider
the test's output as a psql result set anymore, so the "-- ignore" directive
doesn't work anymore. Use the more common "-- start_ignore"/"-- end_ignore"
block instead.

(I'm not sure how useful the test is, if we don't check
the output, but thats a different story.)