Make gptransfer test's output understandable to gpdiff.
The gptransfer behave test was using gpdiff to compare data between the
source and target systems, and was relying on gpdiff to mask row order
differences. However, after 1f44603af2, gpdiff no longer recognized the
results as psql result sets, because it did not echo the SELECT statements
to the output. gpdiff expects to see those. Fix, by echoing the
statements, like in pg_regress. That makes the output, if there are any
differences, more readable anyway.

While we're at it, change the gpdiff invocation to produce a unified diff.
If the test fails, because there is a difference, that makes the output
a lot more readable.