Fix assertion failure in join planning.
cdbpath_motion_for_join() was sometimes returning an incorrect locus for a
join between SingleQE and Hashed loci. This happened, when even the "last
resort" strategy to move hashed side to the single QE failed. This can
happen at least in the query that's added to the regression tests. The
query involves a nested loop join path, when one side is a SingleQE locus
and the other side is a Hashed locus, and there are no join predicates
that can be used to determine the resulting locus.

While we're at it, turn the Assertion that this tripped, and some related
ones at the same place, into elog()s. No need to crash the whole server if
the planner screws up, and it'd be good to perform these sanity checks in
production, too.

The failure of the "last resort" codepath was left unhandled by commit
0522e96074. Fixes https://github.com/greenplum-db/gpdb/issues/6643.

Reviewed-by: Paul Guo <pguo@pivotal.io>