Disable auto segment rebalance during gpstart.
gpstart tried to automatically rebalances the cluster if synced
segment pairs are not in their preferred segment roles (primary or
mirror). This worked and was basically free in file replication. As
part of the cluster start, the gpstart utility would see that the
primary and mirror pair were both in up/sync state but segment roles
reversed in the catalog. It was simple to just send the correct
filerep signals to switch their segment roles to their preferred role.

With WAL replication, this is not as trivial.  The primary and mirror
segments themselves are very aware of their segment roles.  If a
segment found a recovery.conf file in their data directory, it'll
automatically start as a mirror.

So, till this is not properly implemented if decided to still be
supported as part of gpstart, removing the currently broken logic in
gpstart.