Revert commit to estimate cardinality for text columns based on NDV (5a32636)
The commit 5a326369ba926349147f4bb6746ebd92d73ef024 causes underestimation of
cardinality in some cases leading to a sub optimal plan being picked.  Whenever
there are predicates of the form f(a) = b, the optimizer cannot estimate the
cardinality of the filter accurately since it doesn't know how the function
f(a) works. It should revert to using the default estimate in such cases.

Co-authored-by: Chris Hajas <chajas@pivotal.io>