Fix the way tabs are displayed in github
We use editor configurations in ORCA where a TAB indents four positions.
However, when we look at our changes in github.com, we currently see
a tab width of 8. There is a way to tell github.com what our preferred
tab size is, and that is by using EditorConfig, which means placing
a .editorconfig file in the top-level directory in our git project.