Fix Limit, NestLoop and MergeJoin to squelch subnodes.
Previously, these nodes didn't squelch the subnodes after returning the
last row, if the EXEC_FLAG_REWIND option was used. That's not right. It
is not the responsibility of these nodes to shield the subnodes from
squelches. If a subnode doesn't want to free resources eagerly, it needs
to deal with that by itself.

This used to cause deadlocks in the regression tests, in query plans
involving SubPlans with Limit nodes, when using the TCP interconnect.
(The TCP interconnect has shorter send queues, making it more prone to
starvation or deadlock if squelching is missing.) The regression failures
were silenced by commit b7bb54389c, but there might be some other queries
that we just haven't hit yet, where it could still be reproduced.

Reviewed-by: Pengzhou Tang <ptang@pivotal.io>