Fix insertion into inherited AO tables.
When we assign the AO segments to insert to, we do that for all inherited
tables that we might insert to, not just partitions. Otherwise, the insertion
fails with:

ERROR:  append-only table "child" file segment "-1" entry does not exist

or you get an assertion failure if assertions are enabled.

Fixes https://github.com/greenplum-db/gpdb/issues/6068.

Reviewed-by: Ashwin Agrawal <aagrawal@pivotal.io>