Declare cursor for update should handle replicated table too
In 9.0 merge, we add bellow rule for FOR UPDATE:

select for update will lock the whole table, we do it at addRangeTableEntry.
The reason is that gpdb is an MPP database, the result tuples may not be on
the same segment. And for cursor statement, reader gang cannot get Xid to lock
the tuples, so we didn't add a LockRows node for distributed table to avoid
it

this rule should also apply to replicated table.