Update the default value of optimizer_penalize_broadcast_threshold.
This commit sets the default value of the guc optimizer_penalize_broadcast_threshold
to 100000. We have seen a lot of cases where a plan with broadcast was chosen
due to underestimation of  cardinality. In such cases a Redistribute motion
would have been better. So this commit will penalize broadcast when the number
of rows is greater than 100000 so that Redistribute is favored more in this
case. We have tested the change on the perf pipeline and do not see any
regression.

Co-authored-by: Chris Hajas <chajas@pivotal.io>