Use zstandard instead of zlib to compress dispatched plans.
Because it's faster and compresses better. Like with the old zlib code, if
libzstd is not available, fall back to no compression. There was debate on
whether we should fall back to zlib, if that's available, but maintaining
alternative code is not free. Zstandard is so much faster than zlib, that
anyone running a production system really should build with it. A
developer's laptop is a different story, and falling back to no
compression is fine for that.

Discussion: https://groups.google.com/a/greenplum.org/d/msg/gpdb-dev/8w7vRuaqJ6c/uHpOJXwyEgAJ
Reviewed-by: Ashwin Agrawal <aagrawal@pivotal.io>
Reviewed-by: Daniel Gustafsson <dgustafsson@pivotal.io>