Remove flaky pg_basebackup scenario
There was a race condition where it was possible that fault was
unexpectedly triggered by WAL sender object independent of
pg_basebackup being run. We could fix it to be more deternimistic by
incrementing wait for triggered count, but the test as a whole didn't
seem to add much value.

Co-authored-by: Ekta Khanna <ekhanna@pivotal.io>