Remove broken test for an ancient bug.
I looked up this issue in the old JIRA instance:

> MPP-8014: bitmap indexes create entries in gp_distribution_policy
>
> postgres=# \d bar
>       Table "public.bar"
>  Column |  Type   | Modifiers
> --------+---------+-----------
>  i      | integer |
> Distributed by: (i)
>
> postgres=# create index bitmap_idx on bar using bitmap(i);
> CREATE INDEX
> postgres=# select localoid::regclass, * from gp_distribution_policy;
>           localoid          | localoid | attrnums
> ----------------------------+----------+----------
>  bar                        |    16398 | {1}
>  pg_bitmapindex.pg_bm_16415 |    16416 |
> (2 rows)

So the problem was that we created gp_distribution_policy entry for the
auxiliary heap table of the bitmap index. We no longer do that, this bug
was fixed 9 years ago. But the test we have in mpp8014 would not fail,
even if the bug reappeared! Let's remove the test, as it's useless in its
current form. It would be nice to have a proper test for that bug, but it
doesn't seem very likely to re-appear any time soon, so it doesn't seem
worth the effort.

Fixes https://github.com/greenplum-db/gpdb/issues/6315