pg_rewind: Update tests to create separate datadirs for each test (#6689)
Prior to this commit, the test recreated the tmp_check_* directory for
each running test. This would lead to loosing the datadir for the
failing test if it wasn't the last one. This commit, creates a new
directory specific to each test and cleans up artifacts of previous
passing tests

Co-authored-by: David Kimura <dkimura@pivotal.io>
Co-authored-by: Ekta Khanna <ekhanna@pivotal.io>