Remove stale replication slots on  mirrors.
Stale replication slots can exist on mirrors that were once acting as
primaries. In this case restart_lsn is non-zero value used in past
replication slot setup. The stale replication slot will continue to
retain xlog on mirror which is problematic and unnecessary.

This patch drops internal replication slot on startup of mirror.

Co-authored-by: Ashwin Agrawal <aagrawal@pivotal.io>
Co-authored-by: Alexandra Wang <lewang@pivotal.io>