gpstate: show gpexpand status.
Added a new gpstate argument `-x` to show gpexpand status.  The status
information is obtained with the gp_expand_get_status() function,
progress information of the data expansion phase can also be showed.

For `gpstate` or `gpstate -s` a `Cluster Expansion = In Progress`
summary is also displayed if gpexpand is in progress.