Fix up on qual_is_pushdown_safe.
Previously in qual_is_pushdown_safe(), we had this function
qual_is_pushdown_safe_set_operation() to check if the qual contains
references to WindowFunc node.

This is unnecessary because if a qual contains references to WindowFunc
node, its corresponding column would be marked as unsafe that prevents
the qual from being pushed down. So remove the related functions.

Reviewed-by: Heikki Linnakangas <hlinnakangas@pivotal.io>