CI: Remove extra sles11 task input for RC job
For GPDB 6 Beta, only Centos 6/7 need to be passing for the same commit
to be a valid release candidate.

This was originally done in this commit: fa63e7abd6add8b3a1eb8bc9e56bef0f1be1bb5e

But the commit was missing an update to the task yaml for the
Release_Candidate job to accompodate removal of the sles11 input.

Authored-by: Kris Macoskey <kmacoskey@pivotal.io>