Remove GPDB_92_MERGE_FIXME from prepunion.c
The GPDB_92_MERGE_FIXME for whether we need to deep copy or memcopy
suffices in case of subroot can be removed as from the subroot all we
care about is the `parse->rtable`, therefore, creating a deep copy of
it is unnecessary.

This commit also removes the `Assert()` which is valid in Upstream but
for GPDB, since we create a new copy of the subplan if two SubPlans
refer to the same initplan. Therefore, when we try to set references for
subqueryscans in plans with copies of subplans refering to same
initplan, we cannot directly Assert on the RelOptInfo's subplan being
same as the subqueryscan's subplan.

Added a test case for the same, which will ensure we do not merge back
the Assert back from Upstream in future merges.

Co-authored-by: Ekta Khanna <ekhanna@pivotal.io>