Remove BitmapHeapPath and UniquePath check in cdbpath_cost_motion()
For cost estimation of MotionPath node, we calculate the rows as
(subpath->rows * cdbpath_segments) for CdbPathLocus_IsReplicated() which
do not have IndexPath, BitmapHeapPath, UniquePath, and
BitmapAppendOnlyPath (which is completely removed in db51634) in its
subpath. Previously, for the above mentioned node we always calculated
the rows as subpath->rows. The reason why the Paths mentioned above are
special is unknown, the logic has always been there, it used to be in
cdbpath_rows() and was refactored as part of commit b2411b5. Therefore
removing the checks all together, and calculating the rows for all
CdbPathLocus_IsReplicated() to be same. We have already removed
IndexPath as part of the 94_STABLE merge.

With this update, we only see one plan change in ICG.

Co-authored-by: Ekta Khanna <ekhanna@pivotal.io>