Fix up incorrect license statements for module
The gp_sparse_vector module was covered by the relicensing done as
part of the Greenplum open sourcing, but a few mentions of previous
licensing remained in the code. The legal situation of this code has
been reviewed by Pivotal legal and is cleared, so remove incorrect
statements and replace with the standard copyright file headers.

This also cleans up a few comments while at it.

Reviewed-by: Cyrus Wadia
Reviewed-by: Heikki Linnakangas <hlinnakangas@pivotal.io>