gpstate: display 'Copying files' status for primary during full recovery
To make it obvious when a basebackup is in progress (such as with
`gprecoverseg -F`), if any primary's WAL senders are actively backing up
logs to a receiver, `gpstate -s` will now display 'Copying files from
primary' for that primary's Mirror Status. This is true even if a mirror
is currently synced and streaming.

Co-authored-by: Adam Berlin <aberlin@pivotal.io>