Support appendoptimized as alias for AO in reloptions
AO tables are long since Append Optimized and not Append Only, but
the reloptions keyword (and much of the backend nomenclature) is
still "appendonly". This adds support for "appendoptimized" as an
alias for "appendonly", in order to improve the user experience.
The created table, its error messages and the default reloptions
will have "appendonly" set as the reloption as this is only a thin
alias.

  db=# create table t (a integer, b integer) with (appendoptimized=true);
  CREATE TABLE
  db=# \d+ t
                      Append-Only Table "public.t"
   Column |  Type   | Modifiers | Storage | Stats target | Description
  --------+---------+-----------+---------+--------------+-------------
   a      | integer |           | plain   |              |
   b      | integer |           | plain   |              |
  Compression Type: None
  Compression Level: 0
  Block Size: 32768
  Checksum: t
  Distributed by: (a)
  Options: appendonly=true

This commit omits any changes to reflect this in the documentation,
a proper discussion on how to do this needs to happen first.

Reviewed-by: Ashwin Agrawal <aagrawal@pivotal.io>