Fix assertion failure in \di+ and add tests.
The 'translate_columns' array must be larger than the number of columns
in the result set, being passed printQuery(). We had added one column,
"Storage", in GPDB, so we must make the array larger, too.

This is a bit fragile, and would go wrong, if there were any translated
columns after the GPDB-added column. But there isn't, and we don't really
do translation in GPDB, anyway, so this seems good enough.

The Storage column isn't actually interesting for indexes, so omit it
for \di.

Add a bunch of tests. For the \di+ that was hitting the assertion, as well
as \d commands, to test the Storage column.

Fixes github issue https://github.com/greenplum-db/gpdb/issues/6792

Reviewed-by: Melanie Plageman <mplageman@pivotal.io>
Reviewed-by: Jimmy Yih <jyih@pivotal.io>
Reviewed-by: Jesse Zhang <jzhang@pivotal.io>