Remove FIXME, there's nothing to do here.
The point of this FIXME was that the code before the 9.2 merge was
possibly broken, because it was missing this code to get the input slot.
I think it was missing before the 9.2 merge, because of a bungled merge
of commit 7fc0f06221, during the 9.0 merge, but now the code in GPDB
master is identical to upstream, and there's nothing to do. Also,
comparing the 8.2 and 5X_STABLE code, it looks correct in 5X_STABLE, as
well, so there's nothing to do there either.