Don't print WARNING for every temporary file being deleted at recovery.
Before this, you would get warnings like this in the log at crash
recovery, for every temporary file that's deleted:

2019-01-28 20:20:45.702848 EET,,,p7513,th-1570674304,,,,0,,,seg1,,,,,"WARNING","01000","could not open directory ""base/pgsql_tmp/pgsql_tmpslice1_tuplestore5876.0"": No such file or directory",,,,,,,,"pgfnames","pgfnames.c",43,

To fix, backport changes PostgreSQL v11, which added the support for
removing temporary directories in upstream. That is, commit dc6c4c9dc2,
and the follow up commits 561885db0 and eeb3c2df42.

Reviewed-by:  Ashwin Agrawal <aagrawal@pivotal.io>