Capture the correct conjucts in the edges extracted from LOJs
This commit applies to the xforms CXformExpandNAryJoinGreedy &
CXformExpandNAryJoinMinCard. These transforms reorder joins (including
LOJs) to produce a plan with a lower cost. In that process, predicates
need to be rearranged with the reordered tables.

It is important to make sure that resultant LOJs are attached to their
corresponding predicates. To make sure that this happens, edges
extracted from LOJ components must reference both the left & right child
components in the LOJ. The code previous relied on the join condition to
contain colrefs from both the children. However, this is not necessary -
a valid SQL query may contain an LOJ with no colrefs from one or both
children.  In that case, ORCA would either produce a wrong plan or
fail with an assert. See included tests for examples.

This commit fixes this issue.