Remove work arounds for SSL_get_current_compression.
gpcloud uses OpenSSL's libcrypto, even if you ran configure
--without-openssl. The #include <openssl/ssl.h> in gpcloud clashed with
the #define in port.h. I suspect the "ssl.h" was a typo, and should've been
"sha.h", because gpcloud only uses OpenSSL for the hash functions. Change
it that way.

It's a bit bogus that it builds with libcrypto, even if you specified no
OpenSSL support in configure, but

Reviewed-by: Daniel Gustafsson <dgustafsson@pivotal.io>