Remove fixme comment as it is not relative to the merge itself
However, it correctly identifies that xidWarnLimit should not be
configurable. The same should also apply for xidStopLimit. The GUCs
for those were added in time immemorial, i.e. significantly before
greenplum was opensourced, with a commit message clearly identifying
their addition as a one shot hotfix.

A proposal for their depracation has been made in the forum.

Co-authored-by: Daniel Gustafsson <dgustafsson@pivotal.io>
Reviewed-by: Heikki Linnakangas <hlinnakangas@pivotal.io>