Drop ivy third-party extensions tarball for centos
All libraries and programs from the third-party extensions tarball are
now provided by the operating system, so we no longer need to pull this
artifact from ivy.

Dependencies that were dropped:
 - bzip2
 - libedit
 - libevent
 - libxml2
 - libyaml
 - liblber (left behind in the tarball by accident)

Co-authored-by: Karen Huddleston <khuddleston@pivotal.io>
Co-authored-by: Amil Khanzada <akhanzada@pivotal.io>

Remove libyaml and liblber libraries from Makefile for centos

We are not providing these libraries from Ivy in rhel6_x86_64 anymore so
they will not be in our ext directory.

Co-authored-by: Karen Huddleston <khuddleston@pivotal.io>
Co-authored-by: Amil Khanzada <akhanzada@pivotal.io>