Error out on multiple writers in CTE
While Greenplum can plan a CTE query with multiple writable expressions,
it cannot execute it as there is a limitation on using a single writer
gang. Until we can support multiple writer gangs, let's error out with
a graceful error message rather than failing during exeucution with a
more cryptic internal error.

Ideally this will be reverted in GPDB 7.X but right now it's much too
close to release for attacking this.

Reviewed-by: Heikki Linnakangas <hlinnakangas@pivotal.io>