Removed FIXME comment from test suite.
We left the comment explaining the difference between GPDB
and upstream, as it will be useful for future merge work.

pg_event_trigger_dropped_objects should observe events run
on the QD when tables are dropped and have the same behavior
as upstream.

There's a possible race-condition because the query no longer
joins, so there's a gap between the lookup of dropped objects
and the query for undroppable objects, but because theres no
other queries using this table or function, it is not a concern.

Co-authored-by: Georgios Kokolatos <gkokolatos@pivotal.io>