gpcheckcat: fix gp_distribution_policy.distkey comparisons
Follow-up to 69ec6926. distkey is no longer a nullable column, so
comparisons against NULL (to check for random distribution) need to be
replaced with checks for an empty int2vector.

Although this change causes the currently failing master tests to pass,
I'm not sure if

    distkey = ''

is really an idiomatic way to check for an empty vector. I expect a
follow-up to this commit.