Change formatting of AO checksums in error messages.
The gpdiff rule in 'ao_checksum_corruption' test assumed that the
checksums were 8 characters wide. That was not always true, however,
because the checksums were not padded with zeros. Padding them with zeros
seems nicer, so change the error messages to do that.

This should fix these buildfarm failures we've been seeing recently:

-ERROR:  header checksum does not match, expected 0xXXXXXXXX and found 0xXXXXXXXX
+ERROR:  header checksum does not match, expected 0x21B733 and found 0x44C333F8

I'm not sure why we started seeing this now, and I didn't see those errors
on my laptop. But it's pure chance whether a checksum happens to begin
with a 0 or not, so it's not that surprising that some completely
unrelated change changed the physical contents of the table. This commit
should make the failure go away.

Reviewed-by: Hubert Zhang <hzhang@pivotal.io>