Hide WITH RECURSIVE under off by defuault GUC
In Greenplum 5.X the recursive CTE feature was hidden behind a GUC as
it wasn't deemed of production quality just yet. Commit 20152cbf6a30e
removed that GUC in order to make stabilization work easier, there are
still enough rough edges to not consider recursive CTE a feature which
is on by default. This brings back the GUC using the same name in order
to be backwards compatible even though "prototype" is a bit misleading
as this point.

Also adds a test and tidies up a few comments in surrounding code.

Discussion: https://groups.google.com/a/greenplum.org/forum/#!topic/gpdb-dev/Lt6l3wDyxTY
Co-authored-by: Georgios Kokolatos <gkokolatos@pivotal.io>