Fix replication slots test (#6858)
Issue is a hardcoded port in the default `net.ipv4.ip_local_port_range`
of 32768 - 60999. It is possible that another processes is using the
port, in which case the test will fail.  This commit changes to a port
outside that range.  Alternatively, We could dynamically fetch free
ports, but that adds extra unnecessary complexity.

Co-authored-by: Alexandra Wang <lewang@pivotal.io>