pg_regress: recurse input/output directory and convert files.
Special code was written for resgroup directory in input and
output. Instead convert it to generic code to recurse if any directory
encountered in input/output directory. This hurdle was faced when
attempting to add test using template framework for walreplication,
where we wished add test in "input/segwalrep" directory. Change in
this commit now enables to add directories easily to input and output.

Since no need exist now for special resgroup directory argument,
removing the same. Given the number of files in resgroup directory is
very less and not expected to grow super large, its okay to generate
the .sql and .out files always along with generating other files.