Fix "cache lookup failed" errors when reshuffling table with domain.
The code to look up the hash functions for a Reshuffle plan used
get_opfamily_proc() instead of the more versatile
cdb_hashproc_in_family() function, which is used in other similar places
where we need to look up the hash functions for a distribution key. Like
in makeCdbHashForRelation(). That lead to errors if the datatype didn't
have a hash function defined directly for the datatype, but only via a
binary-coercible cast. Domain and enum types are such cases, for example.

Fixes https://github.com/greenplum-db/gpdb/issues/6901

Reviewed-by: David Kimura <dkimura@pivotal.io>
Reviewed-by: Venkatesh Raghavan <vraghavan@pivotal.io>