Remove 9.4 fixme comment without any code change
In Greenplum, during vacuuming of an append optimized relation, the
query dispatcher needs to update its tupcount information from the segments.
Vacuum calls UpdateMasterAosegTotals with the appropriate snapshot as an
argument.

UpdateMasterAosegTotals was introduced in <d9415f4cc5db95a0> (closed source)
and it was called with the special static `SnapshotNow`. Possibly due to
oversight, while gathering the information of columnar append optimized
relations, it reused the static instead of the arg, which happened to match.

In 9.4 SnapshotNow was removed in favor of MVCC catalog scans.

The comment wondered if the use of apparently different snapshots was
intentional. The conclusion is that it was not. Vacuum is holding the
appropriate snapshot which is now used in both cases.

Reviewed-by: Heikki Linnakangas <hlinnakangas@pivotal.io>
Reviewed-by: Ashwin Agrawal <aagrawal@pivotal.io>