Fix nonfunctional gpcheckcat constraint Behave test
One of the Behave tests was calling gpcheckcat -R with an invalid
check name, causing the test to silently fail.  This commit corrected
that call.

Additionally, the SQL script used to generate the error for that test
was making a catalog change that would be nearly impossible to hit "in
the wild," and was causing the repair script to encounter an assertion
failure.  The script has been changed to generate a more reasonable
error, one that is more relevant to what we want to test.

Finally, the Behave step for running repair scripts has been modified
to actually error out if a repair script encounters a non-zero return
code (when it was silently failing before) and give the user a
descriptive error message.

Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>