Rename all HLL functions & structures in GPDB
They are renamed to start with a gp_/GP_/Gp prefix (as appropriate).
This will prevent any namespace clashes when building & running external
HLL-related GPDB extensions.

I decided to keep the directory name that same since that doesn't matter
for name conflicts, and none of the other directories in utils started
with the gp_ prefix.