Silence pg_dump warnings about dependency loops on bitmap indexes.
The >= 9.4 version of the query to get all relations was missing the qual
to filter out the auxiliary heaps and indexes of bitmap indexes. The
quals were added in commit 9cafd3fe39, but got lost from the latest query
in the 9.4 merge. Put it back.

Also add missing fields to the 9.1 query, although this is just pro forma,
because there has been no released version of GPDB based on 9.1.

Fixes https://github.com/greenplum-db/gpdb/issues/6900