Support MCV based cardinality estimation for text columns
Prior to this patch, the MCVs was not being used to compute the cardinality for
text based columns. This would cause errors in the cardinality and lead to a
sub optimal plan being picked. This patch enables ORCA to use the MCVs for
equal and not equal operators for text columns.

Co-authored-by: Abhijit Subramanya <asubramanya@pivotal.io>
Co-authored-by: Bhuvnesh Chaudhary <bchaudhary@pivotal.io>