Record pg_basebackup progress during gpconfigurenewsegment runs
When running gprecoverseg, gpinitstandby, or gpaddmirrors, we actually
run gpconfigurenewsegment to execute pg_basebackup.  Log the progress
output of pg_basebackup to a temporary file for user and/or utility
consumption.  The pg_basebackup progress output is logged to a
temporary file located in ~/gpAdminLogs or wherever the user specified
with -l flag used by most Greenplum Python utilities.  The file is
removed after a successful run.  The pg_basebackup output contains
carriage returns so the user must deal with it themselves through
their editor of choice.

Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>
Co-authored-by: Shoaib Lari <slari@pivotal.io>
Co-authored-by: Mark Sliva <msliva@pivotal.io>