Avoid changing rd_cdbpolicy at relcache invalidation.
The code in CTranslatorQueryToDXL::NoteDistributionPolicyOpclasses() would
sometimes crash or throw an error that leads to falling back to postgres
planner, if a relcache invalidation happened after fetching the
Relation->rd_cdbpolicy to a local variable. To fix, change relcache
invalidation to not change rd_cdbpolicy, if there has been no actual
change to the relation's policy. This is similar to how rd_rel is handled.

Since we're holding a lock on the table, the policy cannot change while
we're looking at it. So this doesn't just reduce the odds; it fixes the
issue for real.

In checkPolicyForUniqueIndex(), allocate rd_cdbpolicy in correct memory
context. All other places where we call GpPolicyReplace() and replcate the
rd_cdbpolicy field in the relcache entry got this right. I think we got
away with this until now, because a relcache invalidation happened soon
afterwards in any command that called this function, which caused the
rd_cdbpolicy in the wrong context to be free'd and re-allocated in the
correct one. Until the previous commit that changed relcache invalidation
to not do that anymore.

Debugged-by: Ekta Khanna <ekhanna@pivotal.io>
Reviewed-by: Ashwin Agrawal <aagrawal@pivotal.io>