Reduce time and stablilize uao_crash_compaction_column test.
Test uao_crash_compaction_column needs to wait for record to be
replayed on mirror for checking. Depending on work load generated by
the tests this can take long time. Hence, moving the test to run at
start of schedule to reduce waiting for that time.

Also, to reduce the flakiness seen in failure increase the number of
retries in test to check if replay_location = flush_location on mirror
from 1000 to 5000.

Co-authored-by: Alexandra Wang <lewang@pivotal.io>