Make sure missing_xlog test leaves the cluster in good state.
missing_xlog brings the mirror down and fiddles with its state. Before
this commit it only validates mirror is up but misses to make its in
sync state. Tests following to this test make assumption of starting
when all the primaries and mirrors are in sync state. Hence, every
test should end leaving primaries and mirrors as up and sync.

Co-authored-by: Alexandra Wang <lewang@pivotal.io>