Add script to fix mdp files
This script will automatically fix mdp files by replacing the query plan
and the plan size. Additionally, it makes other errors easier to read.

The script can run under 2 modes. The first, you pass in a file containing the
failed tests from the ctest output. This will re-run those specific tests and replace the
appropriate mdp files. This can be run with ./fix_mdps
<failed_tests_file>

The second mode will take in the entire ctest log file and is run with
./fix_mdps <log_file>. This will not re-run ctest and will instead just
use the ctest output to fix the mdp files.

This is an initial commit to get the script in the repo. It's still
rough around the edges and there's likely additional features that would
be nice, but this is a starting point that has proven useful.

Authored-by: Chris Hajas <chajas@pivotal.io>