Fix call to update_mergeclause_eclasses().
update_mergeclause_eclasses() asserts that 'left_ac' and 'right_ac' are
not NULL. Therefore, it makes no sense to call it when left_ac is NULL.
But we should call it unconditionally, like in other similar functions in
the same file.

I can't come up with a test case that would trip the assertion, the ECs
have been correctly set by the time we get here. And I haven't been able
to come up with a query where the lack of the call would matter, although
I think it is potentially a problem. So, just fix it.

Also remove redundant assertions on left_ac and right_ac being not NULL.
update_mergeclause_eclasses() asserts that, too.

Fixes https://github.com/greenplum-db/gpdb/issues/5372

Reviewed-by: Melanie Plageman <mplageman@pivotal.io>