Handle parameterized paths correctly when creating a join path. (#6770)
After we have parameterized path since pg 9.2 and lateral (since pg9.3 although
we do not support the full functionality), merge join path and hash join path
need to consider that. Besides, for nestloop path, the previous code is wrong.

1. It did not allow motion for paths include index (path_contains_inner_index()).
   That is wrong.  Here are two examples of index paths which allow motion.

->  Broadcast Motion 3:3  (slice1; segments: 3)  (cost=0.17..24735.67 rows=86100 width=0)
    ->  Index Only Scan using t2i on t2  (cost=0.17..21291.67 rows=28700 width=0)

->  Broadcast Motion 1:3  (slice1; segments: 1)  (cost=0.17..6205.12 rows=259 width=8)
    ->  Index Scan using t2i on t2  (cost=0.17..6201.67 rows=29 width=8)
        Index Cond: (4 = a)

2. The inner path and outer path might require upper nodes for parameterized
   paths so current code code
     bms_overlap(inner_req_outer, outer_path->parent->relids)
   is definitely not sufficient, besides, outer_path could have paramterized
   paths also.

For nestloop join, case 1 is covered by the test case added in join_gp.
For case 2, the test case in join.sql (although ignored) in this patch
actually partially tested.

Note the change in this patch is conservative. In theory, we could refer
subplan code to allow broadcast for base rel if needed (for this solution
no motion is needed), but that needs much effort and does not seem
to be deserved given we will probably refactor related code for the
lateral support in the near future.