Set VARTAG_ONDISK to the appropriate value for Greenplum
Upstream added support for multiple kinds of external toast datums (commit
3682025015390a). In the heart of the implementation, the representation of the
inline portion of a short varlena was modified to store a tag describing the
location of the datum pointed to, namely memory or disk. The tag took the place
of the member which was describing the length of the datum pointed to. Since in
upstream the length, for on disk datums, was always set to (VARHDRSZ_EXTERNAL +
sizeof(struct varatt_external)), the enum for the corresponding tag was chosen
to be exactly that, i.e. 18.

In Greenplum, the exact same thing happens. However, due to historic reasons,
there is an additional two byte padding in the struct. That representation has
(and still is) been reflected in VARHDRSZ_EXTERNAL, which means that Greenplum
has been storing a value for length two bytes longer than upstream.

This commit updates the value of VARTAG_ONDISK to match the value that Greenplum
has been historically storing. The other solution would have been to re-write
the data during upgrade, but it seems unreasonably risky and evasive to do so.

This commit also removes a comment that stated that Greenplum did not always set
the length of the datum pointed to. Extensive search in the latest 5 and 4
versions of Greenplum did not found this to be true anymore. If it were, then
some data rewriting would have been required while upgrading Greenplum.

Previous versions of Greenplum, (pre-2009) have not been checked and it should
be considered that data from those versions will break if binary upgraded to the
current version.

Removes a GPDB_94_MERGE_FIXME.

Co-authored-by: Daniel Gustafsson <dgustafsson@pivotal.io>

Reviewed-by: Heikki Linnakangas <hlinnakangas@pivotal.io>