gpexpand: always store schema tables in 'postgres' db
gpexpand used to store schema tables in the database specified by the
user with -D option, this caused some complicated handlings, for example
gpexpand needed to scan all the databases to ensure schema tables in the
specified database are correct.

To simplify this we retired the -D option, and always use the 'postgres'
database to store the schema tables.