Fix tests after adding Add AttachPartitionEnsureIndexes
This commit is part of Add Partitioned Indexes #7047.

AttachPartitionEnsureIndexes iterates through all the indexes on an
incoming partition and adds INTERNAL_AUTO dependencies to the ones match
the definition of the parent partition's partitioned indexes.

These are all the tests that broke when we started exchanging/testing
for regular indexes on partitioned tables.

Co-authored-by: Kalen Krempely <kkrempely@pivotal.io>