Enable simple DROP INDEX CONCURRENTLY tests
The enabled section verify only the serialized use case. Proper
concurrent tests exist and are enabled in isolation <160984c8c8>.

The exact test case with upstream could not be maintained,
due to the corresponding create index concurrently use case being
disabled. In order to keep future merge conflicts to a minimum the
tests have been moved to a greenplum specific test file. A new
file was needed because the test case did not semantically fit
the cases in the existing files.

Removes 93 merge fixme.

Reviewed-by: Melanie Plageman <mplageman@pivotal.io>
Reviewed-by: Daniel Gustafsson <dgustafsson@pivotal.io>