gpcheckcat: fix repair for randomly distributed tables with invalid constraints
The two separate queries made as part of the distribution_policy checks
overlapped when the table in question was randomly distributed. This led
to two copies of the same DROP CONSTRAINT statement in the repair
scripts, which failed.

To fix, we can remove the overlapping WHERE clause from the second
query, and only focus on column-distributed tables there.

Co-authored-by: Jimmy Yih <jyih@pivotal.io>