Syntax fix to compile libpq on non-C99 compiler (#7033)
fe-misc.c belongs to libpq frontend. libpq is needed by pygresql.
On windows python 2.7 is compiled with VS 2008, and so must be all
extensions including pygresql. VS 2008 is not compliant with C99.
Most part of libpq comes from upstream and is friendly with non-C99
compiler. We only need to fix pqFlushNonBlocking() which is
introduced in commit 510a20b6c2e and unique to GPDB.