explain: fix 'rows' of partial / replicated tables
A replicated table has a full replica of the data on each segment, so
the 'rows' in the EXPLAIN output should not be scaled.

A partial table's 'rows' in the EXPLAIN output should be scaled with the
numsegments of itself.

We used to scale both of above cases with the cluster size in the
EXPLAIN output, so the 'rows' were incorrectly displayed.  It's only a
bug in EXPLAIN output, the cost calculation of the plan is not affected.