Reuse the identifier when a QE is destroyed
Now each QE in a session is assigned with a unique identifier, meanwhile, QD
dispatches a slice table to all QEs and each slice in the slice table has a
bitmapset of QE identifiers. QE go through all slices and decide the slice it
belongs to by checking its identifier in the bitmapset.

A problem is, the QE identifier was never resued when a QE was destroyed, so the
identifier was incrementally increased until the bitmapset is inefficient and
insufficient to hold it.

This commit fixes it by reusing the QE identifiers to limit it in a reasonable
range.

Co-authored-by: Ashwin Agrawal <aagrawal@pivotal.io>