ItemPointer of AO tables need to be fixed
We generate the fake ItemPointer for AO table's tuple, but it is
a little different from the HEAP table's ItemPointer, the offset
of the pointer can be 0 when it is a fake ItemPointer, but we
treat 0 to be invalid in the previous codes, In order to run
correctly, we set the 16th bit of offset to be 1 as a flag, then
it can pass some check, but the value is not the real value, so when
using the fake ItemPointer, we have to note to process the flag.

The bitmap index build code expects to see the tuple IDs in order. 
It gets confused when it sees offset number 32768(0x8000) 
before offset number 1, this commit convert 32768(0x8000) to
be 0.