gpcheckcat: handle partition distclass mismatches and don't repair
After the changes to gp_distribution_policy to add distribution opclass,
the distribution policy checks in gpcheckcat need to be updated to
consider the distclass column as well.

Additionally, the old repair action that was taken as part of this check
is no longer valid, since we can't redistribute a child partition
directly. Instead, leave the correct action up to the DB admin.

Co-authored-by: Jimmy Yih <jyih@pivotal.io>
Co-authored-by: Taylor Vesely <tvesely@pivotal.io>