Add support for Squelching a ModifyTable node
An inner ModifyTable node must run to completion even if the outer
node can be satisfied with a squelched inner. Ensure to run the node
to completion when asked to Squelch to not risk losing modifications.

This adds a testcase from the upstream with test suite to the GPDB
with_clause suite. The original test is under an ignore block, but
even with lifting that the output is different due to state being
set up by prior tests which happen to fail in GPDB.

Reviewed-by: Heikki Linnakangas <hlinnakangas@pivotal.io>