Break up the ignore block in the with suite
A large set of tests were wrapped in an ignore block in the with
suite due to them not working properly in the past. Since most of
these have been addressed, it's time to break up the block and
ensure testing coverage.

This removes as much of the ignore as possible, and updates the
underlying returned data to match. This will create merge conflicts
with upstream but since we wont merge more code before cutting the
next release it's better to have sane tests for the lifecycle of
the next release, and we can always revert this on master as we
statrt merging again.

The trigger tests are left under ignore, even though they seem to
work quite well, since atmsort cannot handle that output yet.

Reviewed-by: Heikki Linnakangas <hlinnakangas@pivotal.io>