Re-enable citext as distribution column in test
The citext test was altered in the past to add a distribution column
other than citext. Since we now allow for extension datatypes to be
distribution columns with the new hashing work, remove the workaround
and re-align the test with upstream.

This also updates the testdata from upstream to work in Greenplum.

Reviewed-by: Heikki Linnakangas <hlinnakangas@pivotal.io>