gpinitstandby: rename -F to -S and document it
This reintroduces commit ba3eb5b416, which was reverted in 659f0ee54b.

After the WALrep changes, the previous -F (filespace) option was
co-opted to be the new standby data directory option. This isn't a
particularly obvious association.

Change the option to -S. (-D would have been better, but that's already
in use as a short alias for --debug.) Also document this option in the
official gpinitstandby help.