Squash the three gppc test suites into one.
We had three test modules to test the gppc partner connector library.
All three were the same in the way the test module was built and called, so
we might as well merge them into a single module to simplify things. This
also makes the tests run faster, as we no longer create a separate database
to test each module, but just one. Leaving behind fewer databases should
also speed up the later stages in installcheck-world, like gpcheckcat,
gp_replica_check, and pg_upgrade, run faster, because we leave behind fewer
databases for them to process.

There was a lot of identical test code and tests among the three modules.
Remove the duplication while we're at it.

The byteafunc2 test function was intentionally changing the pass-by-ref
argument, which is a big no-no. It doesn't make much sense to test that,
because that's a programmer error, and all bets are off. Change the test to
do something more interesting.