Rename recursive CTE guc to remove _prototype
The GUC which enables recursive CTEs is in the currently released
version called gp_recursive_cte_prototype, but in order to reflect
the current state of the code it's now renamed to gp_recursive_cte.
By default the GUC is still off, but that might change before we
ship the next release.

The previous GUC name is still supported, but marked as deprecated,
in order to make upgrades easier.

Reviewed-by: Ivan Novick <inovick@pivotal.io>
Reviewed-by: Georgios Kokolatos <gkokolatos@pivotal.io>